requirements.txt
# Ansible and dependencies
ansible>=2.10.0,<3.0.0
ansible-core>=2.12.0
ansible-lint>=6.0.0

# GitHub API interaction
PyGithub>=2.0.0
requests>=2.28.0

# JWT for GitHub App authentication
PyJWT>=2.8.0
cryptography>=41.0.0

# Testing
pytest>=7.4.0
pytest-ansible>=3.1.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# YAML handling
PyYAML>=6.0

# Jinja2 for templating
Jinja2>=3.1.0

# Utilities
python-dotenv>=1.0.0
